backend/requirements.txt
fastapi==0.70.0
pandas==1.3.4
pandas_stubs==1.2.0.1
python_dateutil==2.8.2
spacy==3.1.3
textract==1.6.3
